@@ -1,10 +1,13 @@
 ﻿using System.Collections.Generic;
 
-namespace _99Bottles
+namespace _99Bottles;
+
+public static class StringExtensions
 {
-    public static class StringExtensions
+    public static string Join(
+        this IEnumerable<string> source,
+        string separator)
     {
-        public static string Join(this IEnumerable<string> source, string separator) =>
-            string.Join(separator: separator, values: source);
+        return string.Join(separator: separator, values: source);
     }
 }
